+++ torchmetrics/regression/mean_absolute_percentage_error.py
@@ -24,9 +24,50 @@ from torchmetrics.metric import Metric
 
 
 class MeanAbsolutePercentageError(Metric):
+    r"""
+    Computes `mean absolute percentage error <https://en.wikipedia.org/wiki/Mean_absolute_percentage_error>`_ (MAPE):
+
+    .. math:: \text{MAPE} = \frac{1}{n}\sum_1^n\frac{|   y_i - \hat{y_i} |}{\max(\epsilon, y_i)}
+
+    Where :math:`y` is a tensor of target values, and :math:`\hat{y}` is a tensor of predictions.
+
+    Args:
+        epsilon:
+            an arbitrary small yet strictly positive number to avoid undefined results when y is zero.
+            default: 1.17e-07. For more information, check the note.
+        compute_on_step:
+            Forward only calls ``update()`` and return None if this is set to False. default: True
+        dist_sync_on_step:
+            Synchronize metric state across processes at each ``forward()``
+            before returning the value at the step. default: False
+        process_group:
+            Specify the process group on which synchronization is called. default: None (which selects the entire world)
+
+    Note:
+        The epsilon value is taken from `scikit-learn's implementation
+        <https://github.com/scikit-learn/scikit-learn/blob/15a949460/sklearn/metrics/_regression.py#L197>`_.
+
+    Note:
+        MAPE output is a non-negative floating point. Best result is 0.0 . But it is important to note that,
+        bad predictions, can lead to arbitarily large values. Especially when some ``target`` values are close to 0.
+        This implementation returns a very large number instead of ``inf``.
+        For more information,  `read here
+        <https://scikit-learn.org/stable/modules/generated/sklearn.metrics.mean_absolute_percentage_error.html>`_.
+
+    Example:
+        >>> from torchmetrics import MeanAbsolutePercentageError
+        >>> target = torch.tensor([1, 10, 1e6])
+        >>> preds = torch.tensor([0.9, 15, 1.2e6])
+        >>> mean_abs_percentage_error = MeanAbsolutePercentageError()
+        >>> mean_abs_percentage_error(preds, target)
+        tensor(0.2667)
+
+
+    """
+
     def __init__(
         self,
-        eps: float = 1.17e-07,
+        epsilon: float = 1.17e-07,
         compute_on_step: bool = True,
         dist_sync_on_step: bool = False,
         process_group: Optional[Any] = None,
@@ -41,16 +82,25 @@ class MeanAbsolutePercentageError(Metric):
 
         self.add_state("sum_abs_per_error", default=tensor(0.0), dist_reduce_fx="sum")
         self.add_state("total", default=tensor(0.0), dist_reduce_fx="sum")
-        self.eps = torch.tensor(eps)
+        self.epsilon = torch.tensor(epsilon)
 
     def update(self, preds: Tensor, target: Tensor) -> None:
+        """
+        Update state with predictions and targets.
 
-        sum_abs_per_error, num_obs = _mean_absolute_percentage_error_update(preds, target, eps=self.eps)
+        Args:
+            preds: Predictions from model
+            target: Ground truth values
+        """
+        sum_abs_per_error, num_obs = _mean_absolute_percentage_error_update(preds, target, epsilon=self.epsilon)
 
         self.sum_abs_per_error += sum_abs_per_error
         self.total += num_obs
 
     def compute(self) -> Tensor:
+        """
+        Computes mean absolute percentage error over state.
+        """
         return _mean_absolute_percentage_error_compute(self.sum_abs_per_error, self.total)
 
     @property